feat: 评分规则系统重构与全局配置管理
- 新增评分规则模型(scoring_rule)与系统配置模型(sys_config)
- 新增评分规则服务、配置服务、内存监控中间件与增强缓存
- 重构Build图书馆初始化、WvW评分、考勤与日志解析服务
- 新增__init__.py完善包结构，补充GW2.txt与builds初始数据
- 移除废弃的WvW报告、EI报告及相关诊断/审计脚本
- 修复Docker entrypoint日志级别大小写兼容问题
- 更新Dockerfile、docker-compose、Nginx配置与部署脚本
- 前端新增评分规则设置页面、服务与相关组件更新
- 更新API端点、类型定义与常量配置

+++ backend/app/models/zevtc_data.py
@@ -28,9 +28,9 @@ class EiPlayer(Base):
 
     __tablename__ = "ei_player"
 
-    player_id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键ID")
+    player_id = Column(Integer, primary_key=True, autoincrement=True, comment="主键ID")
     log_id = Column(
-        BigInteger, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
+        Integer, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
         comment="关联日志ID"
     )
     agent_index = Column(Integer, comment="关联evtc_agent索引")
@@ -61,9 +61,9 @@ class EiTarget(Base):
 
     __tablename__ = "ei_target"
 
-    target_id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键ID")
+    target_id = Column(Integer, primary_key=True, autoincrement=True, comment="主键ID")
     log_id = Column(
-        BigInteger, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
+        Integer, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
         comment="关联日志ID"
     )
     agent_index = Column(Integer, comment="关联evtc_agent索引")
@@ -85,9 +85,9 @@ class EiSkillMap(Base):
 
     __tablename__ = "ei_skill_map"
 
-    map_id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键ID")
+    map_id = Column(Integer, primary_key=True, autoincrement=True, comment="主键ID")
     log_id = Column(
-        BigInteger, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
+        Integer, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
         comment="关联日志ID"
     )
     skill_key = Column(String(20), nullable=False, comment="JSON中的技能键(s12345)")
@@ -108,9 +108,9 @@ class EiPhase(Base):
 
     __tablename__ = "ei_phase"
 
-    phase_id = Column(BigInteger, primary_key=True, autoincrement=True, comment="主键ID")
+    phase_id = Column(Integer, primary_key=True, autoincrement=True, comment="主键ID")
     log_id = Column(
-        BigInteger, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
+        Integer, ForeignKey("evtc_log.log_id", ondelete="CASCADE"), nullable=False,
         comment="关联日志ID"
     )
     phase_index = Column(Integer, nullable=False, comment="阶段索引")
